requirements.txt
filelock==3.16.1
gitdb==4.0.11
GitPython==3.1.43
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
python-dotenv==1.0.1
smmap==5.0.1
watchdog==6.0.0
